More timeline progress
Added frame selecting to the timeline. Now it feels a lot more like an
actual timeline and a lot less like a scrollable picture. Also deleted
the canvas manager class because idfk what i was doin with that. As
well, added a savable interface for use later when I want to make saving
projects.