StickEditor update, rectangle fixes.
Put in the stick editor; It works and all. There are still a couple of
bugs to fix, one being that a runtime error is thrown when you try to
add a custom figure. I also fixed the line not drawing on rectangles,
the fill draws, it just defaults to white for some reason.