Added function to read stickcustom from stream
CustomFigSaver.saveFigure(BinaryReader, StickCustom) was added last
commit, however I forgot to include a function to read from a stream.

CustomFigLoader.loadStickFile(BinaryReader). Returns a StickCustom built
from the filestream. Be sure that the position of the BinaryReader is at
the very start of the .tzf file. Of course, this will advance the stream
the length of a .tzf file.